poc_ai_sast/requirements.txt
semgrep>=1.60.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
jinja2>=3.1.0
httpx>=0.27.0
rich>=13.0.0
pyyaml>=6.0.0
pytest>=8.0.0
openai>=1.0.0
